updates to fix issues in pull request
